feat(PPDSC-2555): rename tools component folde
