+++ site/components/svg-previewer/__tests__/expected-file-mock.tsx
