+++ site/components/svg-previewer/__tests__/figma-svg-mock-data.tsx
